+++ mapbox_gl_platform_interface/lib/src/mapbox_gl_platform_interface.dart
@@ -66,8 +66,9 @@ abstract class MapboxGlPlatform {
       ArgumentCallbacks<void>();
 
   final ArgumentCallbacks<void> onMapIdlePlatform = ArgumentCallbacks<void>();
-  
-  final ArgumentCallbacks<UserLocation> onUserLocationUpdatedPlatform = ArgumentCallbacks<UserLocation>();
+
+  final ArgumentCallbacks<UserLocation> onUserLocationUpdatedPlatform =
+      ArgumentCallbacks<UserLocation>();
 
   Future<void> initPlatform(int id) async {
     throw UnimplementedError('initPlatform() has not been implemented.');
@@ -119,8 +120,9 @@ abstract class MapboxGlPlatform {
   Future<bool> getTelemetryEnabled() async {
     throw UnimplementedError('getTelemetryEnabled() has not been implemented.');
   }
-  
-  Future<List<Symbol>> addSymbols(List<SymbolOptions> options, [List<Map> data]) async {
+
+  Future<List<Symbol>> addSymbols(List<SymbolOptions> options,
+      [List<Map> data]) async {
     throw UnimplementedError('addSymbols() has not been implemented.');
   }
 
@@ -136,6 +138,11 @@ abstract class MapboxGlPlatform {
     throw UnimplementedError('addLine() has not been implemented.');
   }
 
+  Future<List<Line>> addLines(List<LineOptions> options,
+      [List<Map> data]) async {
+    throw UnimplementedError('addLines() has not been implemented.');
+  }
+
   Future<void> updateLine(Line line, LineOptions changes) async {
     throw UnimplementedError('updateLine() has not been implemented.');
   }
@@ -144,8 +151,17 @@ abstract class MapboxGlPlatform {
     throw UnimplementedError('removeLine() has not been implemented.');
   }
 
+  Future<void> removeLines(Iterable<String> ids) async {
+    throw UnimplementedError('removeLines() has not been implemented.');
+  }
+
   Future<Circle> addCircle(CircleOptions options, [Map data]) async {
     throw UnimplementedError('addCircle() has not been implemented.');
+  }
+
+  Future<List<Circle>> addCircles(List<CircleOptions> options,
+      [List<Map> data]) async {
+    throw UnimplementedError('addCircles() has not been implemented.');
   }
 
   Future<void> updateCircle(Circle circle, CircleOptions changes) async {
@@ -168,16 +184,29 @@ abstract class MapboxGlPlatform {
     throw UnimplementedError('removeCircle() has not been implemented.');
   }
 
+  Future<void> removeCircles(Iterable<String> ids) async {
+    throw UnimplementedError('removeCircles() has not been implemented.');
+  }
+
   Future<Fill> addFill(FillOptions options, [Map data]) async {
     throw UnimplementedError('addFill() has not been implemented.');
   }
 
-  Future<void>updateFill(Fill fill, FillOptions changes) async {
+  Future<List<Fill>> addFills(List<FillOptions> options,
+      [List<Map> data]) async {
+    throw UnimplementedError('addFills() has not been implemented.');
+  }
+
+  Future<void> updateFill(Fill fill, FillOptions changes) async {
     throw UnimplementedError('updateFill() has not been implemented.');
   }
 
   Future<void> removeFill(String fillId) async {
     throw UnimplementedError('removeFill() has not been implemented.');
+  }
+
+  Future<void> removeFills(Iterable<String> fillIds) async {
+    throw UnimplementedError('removeFills() has not been implemented.');
   }
 
   Future<List> queryRenderedFeatures(
@@ -231,8 +260,8 @@ abstract class MapboxGlPlatform {
         'setSymbolTextIgnorePlacement() has not been implemented.');
   }
 
-  Future<void> addImageSource(String imageSourceId, Uint8List bytes,
-      LatLngQuad coordinates) async {
+  Future<void> addImageSource(
+      String imageSourceId, Uint8List bytes, LatLngQuad coordinates) async {
     throw UnimplementedError('addImageSource() has not been implemented.');
   }
 
@@ -244,7 +273,8 @@ abstract class MapboxGlPlatform {
     throw UnimplementedError('addLayer() has not been implemented.');
   }
 
-  Future<void> addLayerBelow(String imageLayerId, String imageSourceId, String belowLayerId) async {
+  Future<void> addLayerBelow(
+      String imageLayerId, String imageSourceId, String belowLayerId) async {
     throw UnimplementedError('addLayerBelow() has not been implemented.');
   }
 
@@ -252,9 +282,8 @@ abstract class MapboxGlPlatform {
     throw UnimplementedError('removeLayer() has not been implemented.');
   }
 
-  Future<Point> toScreenLocation(LatLng latLng) async{
-    throw UnimplementedError(
-        'toScreenLocation() has not been implemented.');
+  Future<Point> toScreenLocation(LatLng latLng) async {
+    throw UnimplementedError('toScreenLocation() has not been implemented.');
   }
 
   Future<List<Point>> toScreenLocationBatch(Iterable<LatLng> latLngs) async{
@@ -267,7 +296,7 @@ abstract class MapboxGlPlatform {
         'toLatLng() has not been implemented.');
   }
 
-  Future<double> getMetersPerPixelAtLatitude(double latitude) async{
+  Future<double> getMetersPerPixelAtLatitude(double latitude) async {
     throw UnimplementedError(
         'getMetersPerPixelAtLatitude() has not been implemented.');
   }